+++ src/segger/cli/segment.py
@@ -1,0 +1,376 @@
+from cyclopts import App, Parameter, Group, validators
+from typing import Annotated, Literal
+from pathlib import Path
+
+from .registry import ParameterRegistry
+
+# Register defaults and descriptions from files directly
+# This is to avoid needing to import all requirements before calling CLI
+registry = ParameterRegistry(framework='cyclopts')
+base_dir = Path(__file__).parent.parent
+to_register = [
+    ("data/data_module.py", "ISTDataModule"),
+    ("data/writer.py", "ISTSegmentationWriter"),
+    ("models/lightning_model.py", "LitISTEncoder"),
+]
+for file_path, class_name in to_register:
+    registry.register_from_file(base_dir / file_path, class_name)
+
+# Parameter groups
+group_io = Group(
+    name="I/O",
+    help="Related to file inputs/outputs.",
+    sort_key=0,
+)
+group_nodes = Group(
+    name="Node Representation",
+    help="Related to transcript and cell node representations.",
+    sort_key=2,
+)
+group_transcripts_graph = Group(
+    name="Transcript-Transcript Graph",
+    help="Related to transcript-transcript graph parameters.",
+    sort_key=3,
+)
+group_prediction = Group(
+    name="Segmentation (Prediction) Graph",
+    help="Related to segmentation prediction graph parameters.",
+    sort_key=4,
+)
+group_tiling = Group(
+    name="Tiling",
+    help="Related to tiling parameters.",
+    sort_key=5,
+)
+group_model = Group(
+    name="Model",
+    help="Related to model architecture and training parameters.",
+    sort_key=6,
+)
+group_loss = Group(
+    name="Loss",
+    help="Related to loss function parameters.",
+    sort_key=7,
+)
+
+app_segment = App(name="segment", help="Run cell segmentation on spatial transcriptomics data.")
+
+@app_segment.command(name="segment")
+def segment(
+    # I/O
+    input_directory: Annotated[Path, registry.get_parameter(
+        "input_directory",
+        alias="-i",
+        group=group_io,
+        validator=validators.Path(exists=True, dir_okay=True),
+    )] = registry.get_default("input_directory"),
+
+    output_directory: Annotated[Path, registry.get_parameter(
+        "output_directory",
+        alias="-o",
+        group=group_io,
+        validator=validators.Path(exists=True, dir_okay=True),
+    )] = registry.get_default("output_directory"),
+    
+
+    # Cell Representation
+    node_representation_dim: Annotated[int, Parameter(
+        help="Number of dimensions used to represent each node type.",
+        validator=validators.Number(gt=0),
+        group=group_nodes,
+        required=False,
+    )] = registry.get_default("cells_embedding_size"),
+
+    cells_representation: Annotated[Literal['pca', 'morphology'], registry.get_parameter(
+        "cells_representation_mode",
+        group=group_nodes,
+    )] = registry.get_default("cells_representation_mode"),
+
+    cells_min_counts: Annotated[int, registry.get_parameter(
+        "cells_min_counts",
+        validator=validators.Number(gte=0),
+        group=group_nodes,
+    )] = registry.get_default("cells_min_counts"),
+
+    cells_clusters_n_neighbors: Annotated[int, registry.get_parameter(
+        "cells_clusters_n_neighbors",
+        validator=validators.Number(gt=0),
+        group=group_nodes,
+    )] = registry.get_default("cells_clusters_n_neighbors"),
+
+    cells_clusters_resolution: Annotated[float, registry.get_parameter(
+        "cells_clusters_resolution",
+        validator=validators.Number(gt=0, lte=5),
+        group=group_nodes,
+    )] = registry.get_default("cells_clusters_resolution"),
+
+
+    # Gene Representation
+    genes_clusters_n_neighbors: Annotated[int, registry.get_parameter(
+        "genes_clusters_n_neighbors",
+        validator=validators.Number(gt=0),
+        group=group_nodes,
+    )] = registry.get_default("genes_clusters_n_neighbors"),
+
+    genes_clusters_resolution: Annotated[float, registry.get_parameter(
+        "genes_clusters_resolution",
+        validator=validators.Number(gt=0, lte=5),
+        group=group_nodes,
+    )] = registry.get_default("genes_clusters_resolution"),
+
+
+    # Transcript-Transcript Graph
+    transcripts_max_k: Annotated[int, registry.get_parameter(
+        "transcripts_graph_max_k",  
+        validator=validators.Number(gt=0),
+        group=group_transcripts_graph,
+    )] = registry.get_default("transcripts_graph_max_k"),
+
+    transcripts_max_dist: Annotated[float, registry.get_parameter(
+        "transcripts_graph_max_dist",
+        validator=validators.Number(gt=0),
+        group=group_transcripts_graph,
+    )] = registry.get_default("transcripts_graph_max_dist"),
+
+
+    # Segmentation (Prediction) Graph
+    prediction_mode: Annotated[
+        Literal["nucleus", "cell", "uniform"],
+        registry.get_parameter(
+            "prediction_graph_mode",
+            group=group_prediction,
+        )
+    ] = registry.get_default("prediction_graph_mode"),
+
+    prediction_max_k: Annotated[int | None, registry.get_parameter(
+        "prediction_graph_max_k",
+        validator=validators.Number(gt=0),
+        group=group_prediction,
+    )] = registry.get_default("prediction_graph_max_k"),
+
+    prediction_expansion_ratio: Annotated[float | None, registry.get_parameter(
+        "prediction_graph_buffer_ratio",
+        validator=validators.Number(gt=0),
+        group=group_prediction,
+    )] = registry.get_default("prediction_graph_buffer_ratio"),
+
+    # Tiling
+    tiling_margin_training: Annotated[float, registry.get_parameter(
+        "tiling_margin_training",
+        validator=validators.Number(gte=0),
+        group=group_tiling,
+    )] = registry.get_default("tiling_margin_training"),
+
+    tiling_margin_prediction: Annotated[float, registry.get_parameter(
+        "tiling_margin_prediction",
+        validator=validators.Number(gte=0),
+        group=group_tiling,
+    )] = registry.get_default("tiling_margin_prediction"),
+
+    max_nodes_per_tile: Annotated[int, registry.get_parameter(
+        "tiling_nodes_per_tile",
+        validator=validators.Number(gt=0),
+        group=group_tiling,
+    )] = registry.get_default("tiling_nodes_per_tile"),
+
+    max_edges_per_batch: Annotated[int, registry.get_parameter(
+        "edges_per_batch",
+        validator=validators.Number(gt=0),
+        group=group_tiling,
+    )] = registry.get_default("edges_per_batch"),
+
+    # Model
+    n_epochs: Annotated[int, Parameter(
+        validator=validators.Number(gt=0),
+        group=group_model,
+        help="Number of training epochs.",
+    )] = 20,
+
+    n_mid_layers: Annotated[int, registry.get_parameter(
+        "n_mid_layers",
+        validator=validators.Number(gte=0),
+        group=group_model,
+    )] = registry.get_default("n_mid_layers"),
+
+    n_heads: Annotated[int, registry.get_parameter(
+        "n_heads",
+        validator=validators.Number(gt=0),
+        group=group_model,
+    )] = registry.get_default("n_heads"),
+
+    hidden_channels: Annotated[int, registry.get_parameter(
+        "hidden_channels",
+        validator=validators.Number(gt=0),
+        group=group_model,
+    )] = registry.get_default("hidden_channels"),
+
+    out_channels: Annotated[int, registry.get_parameter(
+        "out_channels",
+        validator=validators.Number(gt=0),
+        group=group_model,
+    )] = registry.get_default("out_channels"),
+
+    learning_rate: Annotated[float, registry.get_parameter(
+        "learning_rate",
+        validator=validators.Number(gt=0),
+        group=group_model,
+    )] = registry.get_default("learning_rate"),
+
+    use_positional_embeddings: Annotated[bool, registry.get_parameter(
+        "use_positional_embeddings",
+        group=group_model,
+    )] = registry.get_default("use_positional_embeddings"),
+
+    normalize_embeddings: Annotated[bool, registry.get_parameter(
+        "normalize_embeddings",
+        group=group_model,
+    )] = registry.get_default("normalize_embeddings"),
+
+    # Loss
+    segmentation_loss: Annotated[
+        Literal["triplet", "bce"],
+        registry.get_parameter(
+            "sg_loss_type",
+            group=group_loss,
+        )
+    ] = registry.get_default("sg_loss_type"),
+
+    transcripts_margin: Annotated[float, registry.get_parameter(
+        "tx_margin",
+        validator=validators.Number(gt=0),
+        group=group_loss,
+    )] = registry.get_default("tx_margin"),
+
+    segmentation_margin: Annotated[float, registry.get_parameter(
+        "sg_margin",
+        validator=validators.Number(gt=0),
+        group=group_loss,
+    )] = registry.get_default("sg_margin"),
+
+    transcripts_loss_weight_start: Annotated[float, registry.get_parameter(
+        "tx_weight_start",
+        validator=validators.Number(gte=0),
+        group=group_loss,
+    )] = registry.get_default("tx_weight_start"),
+
+    transcripts_loss_weight_end: Annotated[float, registry.get_parameter(
+        "tx_weight_end",
+        validator=validators.Number(gte=0),
+        group=group_loss,
+    )] = registry.get_default("tx_weight_end"),
+
+    cells_loss_weight_start: Annotated[float, registry.get_parameter(
+        "bd_weight_start",
+        validator=validators.Number(gte=0),
+        group=group_loss,
+    )] = registry.get_default("bd_weight_start"),
+
+    cells_loss_weight_end: Annotated[float, registry.get_parameter(
+        "bd_weight_end",
+        validator=validators.Number(gte=0),
+        group=group_loss,
+    )] = registry.get_default("bd_weight_end"),
+
+    segmentation_loss_weight_start: Annotated[float, registry.get_parameter(
+        "sg_weight_start",
+        validator=validators.Number(gte=0),
+        group=group_loss,
+    )] = registry.get_default("sg_weight_start"),
+
+    segmentation_loss_weight_end: Annotated[float, registry.get_parameter(
+        "sg_weight_end",
+        validator=validators.Number(gte=0),
+        group=group_loss,
+    )] = registry.get_default("sg_weight_end"),
+
+    debug_mode: Annotated[str, Parameter(
+        help="Whether to save additional debug information. If `predict`, saves trainer and predictions. If `all`, saves trainer, predictions, and also intermediate outputs at the end of each epoch (this can consume a lot of disk space).",
+    )] = "none",
+):
+    """Run cell segmentation on spatial transcriptomics data."""
+    # Remove SLURM environment autodetect
+    from lightning.pytorch.plugins.environments import SLURMEnvironment
+    SLURMEnvironment.detect = lambda: False
+
+    # Setup Lightning Data Module
+    from ..data import ISTDataModule
+    datamodule = ISTDataModule(
+        input_directory=input_directory,
+        cells_representation_mode=cells_representation,
+        cells_embedding_size=node_representation_dim,
+        cells_min_counts=cells_min_counts,
+        cells_clusters_n_neighbors=cells_clusters_n_neighbors,
+        cells_clusters_resolution=cells_clusters_resolution,
+        genes_clusters_n_neighbors=genes_clusters_n_neighbors,
+        genes_clusters_resolution=genes_clusters_resolution,
+        transcripts_graph_max_k=transcripts_max_k,
+        transcripts_graph_max_dist=transcripts_max_dist,
+        prediction_graph_mode=prediction_mode,
+        prediction_graph_max_k=prediction_max_k,
+        prediction_graph_buffer_ratio=prediction_expansion_ratio,
+        tiling_margin_training=tiling_margin_training,
+        tiling_margin_prediction=tiling_margin_prediction,
+        tiling_nodes_per_tile=max_nodes_per_tile,
+        edges_per_batch=max_edges_per_batch,
+    )
+    
+    # Setup Lightning Model
+    from ..models import LitISTEncoder
+    n_genes = datamodule.ad.shape[1]
+    model = LitISTEncoder(
+        n_genes=n_genes,
+        n_mid_layers=n_mid_layers,
+        n_heads=n_heads,
+        in_channels=node_representation_dim,
+        hidden_channels=hidden_channels,
+        out_channels=out_channels,
+        learning_rate=learning_rate,
+        sg_loss_type=segmentation_loss,
+        tx_margin=transcripts_margin,
+        sg_margin=segmentation_margin,
+        tx_weight_start=transcripts_loss_weight_start,
+        tx_weight_end=transcripts_loss_weight_end,
+        bd_weight_start=cells_loss_weight_start,
+        bd_weight_end=cells_loss_weight_end,
+        sg_weight_start=segmentation_loss_weight_start,
+        sg_weight_end=segmentation_loss_weight_end,
+        normalize_embeddings=normalize_embeddings,
+        use_positional_embeddings=use_positional_embeddings,
+    )
+
+    # Setup Lightning Trainer
+    from lightning.pytorch.loggers import CSVLogger
+    from ..data import ISTSegmentationWriter
+    from lightning.pytorch import Trainer
+    logger = CSVLogger(output_directory)
+    writer = ISTSegmentationWriter(output_directory)
+    trainer = Trainer(
+        logger=logger,
+        max_epochs=n_epochs,
+        reload_dataloaders_every_n_epochs=1,
+        callbacks=[writer],
+    )
+
+    # Training
+    trainer.fit(model=model, datamodule=datamodule)
+
+    # Prediction
+    predictions = trainer.predict(model=model, datamodule=datamodule)
+
+    # Save intermediaries (trainer, predictions) if debug mode is "predict" or "all"
+    if debug_mode in ("predict", "all"):
+        import pickle
+        debug_dir = output_directory / "debug"
+        debug_dir.mkdir(exist_ok=True)
+        trainer.save_checkpoint(debug_dir / "trainer.ckpt")
+        with open(debug_dir / "predictions.pkl", "wb") as f:
+            pickle.dump(predictions, f)
+
+    writer.write_on_epoch_end(
+        trainer=trainer,
+        pl_module=model,
+        predictions=predictions,
+        batch_indices=[],
+    )
+
+
